% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ah.R
\name{Ah}
\alias{Ah}
\title{Ah}
\usage{
Ah(polygon)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}
}
\value{
A unit sf layer with an parID column
}
\description{
This function draws a Ah pattern, with the following  helper fuction \link[soilprofile2]{basic_point}
}
