% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_geometry.R
\name{basic_line}
\alias{basic_line}
\title{basic line}
\usage{
basic_line(polygon, cellnumber = 22, rotation = 12)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}

\item{cellnumber}{The distance between the lines}

\item{rotation}{The angle of the line}
}
\value{
A unit point layer (sf) with one rown in the parID column
}
\description{
This function draws a line-layer, with the following helper fuction \link[sf]{st_bbox}
}
