% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_geometry.R
\name{basic_point}
\alias{basic_point}
\title{basic point}
\usage{
basic_point(polygon, cellsize = 12, random = FALSE)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}

\item{random}{if random or not}
}
\value{
A unit point layer (sf) with one rown in the parID column
}
\description{
This function draws a pointlayer, with the following helper fuction \link[sf]{st_sample} by random points and \link[sf]{st_make_grid}
}
