% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_polygon.R
\name{basic_polygon}
\alias{basic_polygon}
\title{basic polygon}
\usage{
basic_polygon(polygon, cellsize, square = F)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}
}
\value{
A unit polygon (sf) with an parID column
}
\description{
This function draws a polygon, with the following  helper fuction \link[sf]{st_make_grid}
}
