% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_geometry.R
\name{basic_random_polygon}
\alias{basic_random_polygon}
\title{basic random polygon}
\usage{
basic_random_polygon(polygon, size, number, nSides, sm = T, ...)
}
\arguments{
\item{polygon}{A Simple Feature polygon or  just a "sfc_MULTIPOLYGON" or "sfc_POLYGON" geometry.}

\item{size}{percentage of output area relative to input area}

\item{number}{number of polygons}

\item{nSides}{number of polygon sides}

\item{sm}{smooth by the \link[smoothr]{smooth} function}

\item{...}{is passed to function \link[smoothr]{smooth}}
}
\value{
A unit multipolygon Simple Feature with one parID column
}
\description{
This function draws a set of random polygons.
}
\examples{
poly_example <- sf::st_sfc(sf::st_polygon(list(matrix(c(0,0,10,0,10,10,0,10,0,0),
ncol=2, byrow=TRUE))))

poly_example1 <- basic_random_polygon(polygon = poly_example, 
                                      size = .3, 
                                      number = 30,
                                      nSides = 5, sm = TRUE)
plot(poly_example)
plot(poly_example1$geometry, add = TRUE, col = "brown")
}
