% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_random_pattern.R
\name{build_random_pattern}
\alias{build_random_pattern}
\title{build random texture}
\usage{
build_random_pattern(polygon, ...)
}
\arguments{
\item{polygon}{A sf tibble with a list fuction column and  a list geom column}

\item{...}{A list with a name column that contains as many horizon 
shortcuts as possible and a column with the corresponding functions that create a pattern.}
}
\value{
This function returns a new geometry of the sf-file equal to \link[soilprofile2]{texture} (which is used internally)
}
\description{
The actual task is to join the horizon file with the corresponding functions  and to apply it
}
