% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cat.R
\name{clean_cat}
\alias{clean_cat}
\title{Clip the point layer by horizons}
\usage{
clean_cat(temp, shape)
}
\arguments{
\item{temp}{The point layer, which has been build by the \link[soilprofile2]{smooth_trans} function.}

\item{shape}{A sf multipolygon set which has been biuld by the   \link[soilprofile2]{st_polygon}}
}
\value{
A point layer with the dimension of the orginal horizons
}
\description{
The buffer allows points to lie outside the horizons, these are cut out with this function.
}
