% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_poly.R
\name{convex_poly}
\alias{convex_poly}
\title{Helper function to extract X and Y coordinates and create ellipses}
\usage{
convex_poly(cord, attr_poly, poly_id)
}
\arguments{
\item{cord}{A numeric vector with the point coordinates specifying the position of the polygon.}

\item{attr_poly}{the parameters described in function \link[soilprofile2]{skeleton}}

\item{poly_id}{an id ...}
}
\value{
a matrix with the x and y coordinates and an id for each polygon
}
\description{
This function builds polygons with given area and number of sides. The starting point is one of the corner points of the polygon.
}
