% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cord_setting.R
\name{cord_setting}
\alias{cord_setting}
\title{"Coordinate creator"}
\usage{
cord_setting(database = df, plot_width = 3)
}
\arguments{
\item{database}{A dataframe, which was build by the function \link[soilprofile2]{data_mod}}

\item{plot_width}{A numeric value that determines the relative width of the  representation.}
}
\value{
This function returns a dataframe with four coordinates for each horizon. Additionally a column with the ID, which can be joind to the source data
}
\description{
This is a simple function that, build coordinates for each horizont. So 
With this you can easily build sf files
}
