% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{data_load}
\alias{data_load}
\title{Load example profildata}
\usage{
data_load(Profile = "SG2")
}
\arguments{
\item{Profile}{The name of the Profile. One of the following names can be selected: "P1", "SG2", "SG5", "SG7"}
}
\value{
A dataframe
}
\description{
The package contains example profile data which can be loaded with this function.
}
