% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_horizont_fun.R
\name{edit_horizont_fun}
\alias{edit_horizont_fun}
\title{edit_horizont_fun}
\usage{
edit_horizont_fun(name)
}
\arguments{
\item{name}{the name of the function same to the horizontname}
}
\description{
With this fuction you can edit and save an existing horizont function
}
