% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Example Data from Gianluca Filippa}
\format{An object of class \code{profile.data.frame} of length 17.}
\usage{
example
}
\description{
A dataset containing soil horizons
}
\keyword{datasets}
