% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genY_fun.R
\name{genY_fun}
\alias{genY_fun}
\title{new Y-values}
\usage{
This function is used internally by the \link[soilprofile2]{line_mod} function. The corresponding Y values are created with the similar function \link[soilprofile2]{genX_fun}.
}
\arguments{
\item{Y}{A numeric vector that giving the start and end of the horizont line}

\item{excited}{Logical value specifying whether to include an exclamation
point after the text}
}
\value{
This function returns 12 Y-value. 10 of this are new generatet by the \link[base]{rnorm}  Funktion
}
\description{
generates new Y-values for the upper horizon limit
}
