% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_mod.R
\name{line_mod}
\alias{line_mod}
\title{creates a sf line}
\usage{
line_mod(df_polygon, mat_line, sm = T)
}
\arguments{
\item{df_polygon}{A character string giving the text the function will print}

\item{mat_line}{A character string giving the text the function will print}

\item{sm}{A character string giving the text the function will print}
}
\value{
This function returns a sf-file, which one line for each horizont
}
\description{
This is a simple function that creates from the horizont coordinates a sf line. 
In zukunft müssen die Parameter für die Form der Linie hier einstellbar sein
}
