% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_poly.R
\name{point_2_polygon}
\alias{point_2_polygon}
\title{Helper function to build polygons by points}
\usage{
point_2_polygon(sf_point)
}
\arguments{
\item{sf_point}{A simple feature with at least the following parameter as colnames:}
}
\value{
A sf polygon layer with one line for each input point
}
\description{
This function builds polygons with given area and number of sides. The starting point is one of the corner points of the polygon.
}
