% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_poly.R
\name{poly_fun}
\alias{poly_fun}
\title{Helper function to calculate a polygon by given number of sides and area.}
\usage{
poly_fun(nSides, area)
}
\arguments{
\item{nSides}{A simple feature with at least the following parameter as colnames:}
}
\value{
A sf polygon layer with one line for each input point
}
\description{
This function builds polygons with given area and number of sides. The starting point is one of the corner points of the polygon. 
The area (A) of a convex polygon can be found with the following formula  
\eqn{A = 1/2 * [(x1*y2 + x2*y3 + ... + xn*y1) - (y1*x2 + y2*x3 + ... + yn*x1)]}
}
