% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_layer.R
\name{process_layer}
\alias{process_layer}
\title{process_layer}
\usage{
process_layer(polygon, number = 1, buffer = 1, rotation = 45, ...)
}
\arguments{
\item{polygon}{four coordinates..}

\item{number}{par variable e.g col}

\item{buffer}{value of variable e.g. blue}

\item{rotation}{angle of the pointgrid}

\item{...}{the "id" which defines the layer}
}
\value{
A unit sf layer with an parID column
}
\description{
This function draws an layer with processes
}
