% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{process_symbols}
\alias{process_symbols}
\title{process_symbols}
\format{An object of class \code{data.frame} with 6 rows and 5 columns.}
\usage{
process_symbols
}
\description{
A data set containing a imaginary soilprofile with 3 horizonts.
}
\keyword{datasets}
