% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_geometry.R
\name{random_line}
\alias{random_line}
\title{basic random line}
\usage{
random_line(polygon, number = 800, line_length = 0.8, variation = 1,
  smoothness = 5)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}

\item{number}{Number of polygons}

\item{line_length}{number of polygon sides}

\item{variation}{smooth by the \link[smoothr]{smooth} function with the methode "ksmooth"}

\item{smoothness}{}
}
\value{
A unit multipolygon (sf) with one parID column
}
\description{
This function draws a set of randompolygons.
}
