% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_geometry.R
\name{random_line_pattern}
\alias{random_line_pattern}
\title{build random texture}
\usage{
random_line_pattern(polygon, ...)
}
\arguments{
\item{polygon}{polygon A Simple Feature or  just  a "sfc_MULTIPOLYGON" or "sfc_POLYGON" geometry}

\item{...}{ignored}
}
\value{
This function returns a new Simple File geometry
}
\description{
The actual task to build random pattern if no texture information and no specific horizont function is available and you still want to draw a texture. for example for black and white representations.
}
