% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root}
\alias{root}
\title{creates a root layer modelling by a L-System}
\usage{
root(Plant = 1, skeleton = spoint, start = c(6, 0), horizont = c(1:3),
  smooth = "chaikin")
}
\arguments{
\item{Plant}{a numeric value between 1 and 6 calling the stored L-system}

\item{skeleton}{A layer with the skeleton content as polygon (bei fehlen gibt es noch einen Fehler)}

\item{start}{The start coordinates (x and y) given in a vector}
}
\value{
This function returns a new sf-line geometry with an id column
}
\description{
Hier sollten man noch einige Sachen einbauen:
kein Fehler bei fehlendem skeleton
smooth
clip mit dem ursprünglichen layer
}
