% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root_position}
\alias{root_position}
\title{creates a root layer modelling by a L-System}
\usage{
root_position(root_inp, horizont, spoint, sample_size)
}
\arguments{
\item{Plant}{a numeric value between 1 and 6 calling the stored L-system}
}
\value{
This function returns a new sf-line geometry with an id column
}
\description{
Hier sollten man noch einige Sachen einbauen:
kein Fehler bei fehlendem skeleton
smooth
clip mit dem ursprünglichen layer
}
