% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_horizont_fun.R
\name{save_horizont_fun}
\alias{save_horizont_fun}
\title{save_horizont_fun}
\usage{
save_horizont_fun(fun, name)
}
\arguments{
\item{fun}{A function}

\item{name}{the name of the function same to the horizontname}
}
\value{
A unit sf layer with an parID column
}
\description{
This function save the new horizont function in the global list
}
