% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_par_setting.R
\name{save_par_setting}
\alias{save_par_setting}
\title{save_horizont_fun}
\usage{
save_par_setting(name, variable, value, par_ID)
}
\arguments{
\item{name}{the name of the function same to the horizontname}

\item{fun}{A function}
}
\value{
A unit sf layer with an parID column
}
\description{
This function save the new horizont function in the global list
}
