% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_par_setting.R
\name{save_struct_poly}
\alias{save_struct_poly}
\title{save_struct_poly}
\usage{
save_struct_poly(name, path)
}
\arguments{
\item{name}{name of the function structure}

\item{path}{path of the png}
}
\value{
A unit sf layer
}
\description{
This function stores the structure. Input is a raster (png) and output is a polygon (sf) needed to store the data for the Shiny app
}
