% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_newsplit.R
\name{sf_newsplit}
\alias{sf_newsplit}
\title{wrapper function for shape_mod}
\usage{
sf_newsplit(shape_areas, mat_line, line_sm = T, Line = F)
}
\arguments{
\item{shape_areas}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}
}
\value{
A unit sf layer with an parID column
}
\description{
This function draws a Ah pattern, with the following  helper fuction \link[soilprofile2]{basic_point}
}
