% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_polgon.R
\name{sf_polgon}
\alias{sf_polgon}
\title{sf polygon builder}
\usage{
sf_polgon(database, df_polygon)
}
\arguments{
\item{database}{The original table with all variables}

\item{df_polygon}{A long table with the columns name, x coordinate and y coordinate for each horizont. Build by the function \link[soilprofile2]{cord_setting}}
}
\value{
This function returns an set of polygons as simple features. Each polygon represents a horizon defined by the name
}
\description{
merge coordinats with the input dataframe by name
}
