% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_soil.R
\name{sf_soil}
\alias{sf_soil}
\title{wrapper function for shape_mod}
\usage{
sf_soil(data, plot_width = 3, line_sm = T, Line = F)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}
}
\value{
A unit sf layer with an parID column
}
\description{
This function draws a Ah pattern, with the following  helper fuction \link[soilprofile2]{basic_point}
}
