% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_trans.R
\name{smooth_trans}
\alias{smooth_trans}
\title{Horizons transition}
\usage{
smooth_trans(Line, database = df, shape_mod = shape_mod, shape = 5,
  seed = 33)
}
\arguments{
\item{Line}{The line layer, which has been build by the \link[soilprofile2]{line_mod} function.}

\item{database}{The original table with the variables}

\item{shape_mod}{The table modified by new lines  \link[soilprofile2]{split_polygon}}
}
\value{
A point layer with a polygons for the color transition
}
\description{
This funktion modifide th horizons transitions. With this function you get waves, tongues and points as transition between the horizons.
}
