% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_example_sf}
\alias{soil_example_sf}
\title{Soilprofile example}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 15 columns.}
\usage{
soil_example_sf
}
\description{
A data set containing a imaginary soilprofile with 3 horizonts.
}
\keyword{datasets}
