% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_default.R
\name{soil_legend}
\alias{soil_legend}
\title{soil_legend}
\usage{
soil_legend(df_legend)
}
\arguments{
\item{df_legend}{A data frame with aesthetics for the legend. In the simplest case the table ("process_symbols") stored in the folder data.}
}
\value{
ggplot
}
\description{
a simple function to visualize the legend in an own "ggplot" plot.
}
\examples{
data("process_symbols")
soil_legend(process_symbols)

}
