% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_default.R
\name{soil_theme}
\alias{soil_theme}
\title{soil_theme}
\usage{
soil_theme()
}
\value{
ggplot themes
}
\description{
a function to call the theme settings of ggplot. The most important is the absence of the X-axis.
}
