% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_polygon.R
\name{split_polygon}
\alias{split_polygon}
\title{Split the soil body by lines to sepreate the horizons}
\usage{
split_polygon(Polygon, Line)
}
\arguments{
\item{Polygon}{The defult soil polygon, which has been build by the \link[soilprofile2]{sf_polygon} function.}

\item{Line}{The modifide  horizont trassition, which has been build by the \link[soilprofile2]{line_mod} function.}
}
\value{
A new sf-Polygon set with modified horizons transitions
}
\description{
If you don't want a straight line as horizon transition, you can create different transition shapes with this function
}
