% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{struc_poly}
\alias{struc_poly}
\title{structure polygons}
\format{Simple feature collection with ...}
\usage{
struc_poly
}
\description{
structure polygons (sf) needed for the Shiny app
}
\keyword{datasets}
