% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture.R
\name{texture}
\alias{texture}
\title{creates a pattern with a given layer and function}
\usage{
texture(shape, fun_horizont, buffer = -1, ...)
}
\arguments{
\item{shape}{A sf tibble with a list fuction column and  a list geom column}

\item{fun_horizont}{The Pattern function, i.e. the column from the "shape" file (könnte man sich vlt auch sparen)}

\item{buffer}{A negetive numeric value that defines the distance to the outer boundary. The default is -0.5}
}
\value{
This function returns a new geometry of the sf-file
}
\description{
For each horizon name there should be a function which creates a corresponding pattern. This function applies the pattern function and creates a new layer. This is smaller than the original layer due to a defined buffer. This is displayed using a line.
}
