% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_random_polygon.R
\name{basic_random_polygon}
\alias{basic_random_polygon}
\title{basic random polygon}
\usage{
basic_random_polygon(polygon, size, number, nSides, sm = T)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}

\item{size}{Percentage area of input polygon}

\item{number}{Number of polygons}

\item{nSides}{number of polygon sides}

\item{sm}{smooth by the \link[smoothr]{smooth} function with the methode "ksmooth"}
}
\value{
A unit multipolygon (sf) with one parID column
}
\description{
This function draws a set of randompolygons.
}
