% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_geometry.R
\name{Bv}
\alias{Bv}
\title{Bv}
\usage{
Bv(polygon, ...)
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}

\item{...}{ignored}
}
\value{
A unit sf layer with an parID column
}
\description{
This function draws a Ah pattern, with the following  helper fuction \link[soilprofile2]{basic_regular_point}
}
