% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_poly.R
\name{convex_poly}
\alias{convex_poly}
\title{Helper function from \link[soilprofile2]{smoth_trans}}
\usage{
convex_poly(data_skel)
}
\arguments{
\item{nSides}{The number of Sides}

\item{area}{The area size}

\item{xstart}{startet X Coordinats,  see \link[soilprofile2]{smoth_trans}}

\item{ystart}{startet Y Coordinats,  see \link[soilprofile2]{smoth_trans}}
}
\value{
A sl polygon layer with one line for each input point
}
\description{
This function builds polygons with given area and number of sides, with an given start point
The area of a convex polygon can be found as: 
deqn{ A = 1/2 * [(x1*y2 + x2*y3 + ... + xn*y1) - (y1*x2 + y2*x3 + ... + yn*x1)]}
}
