% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_mod.R
\name{expand_x}
\alias{expand_x}
\title{new X-values}
\usage{
This function is used internally by the \link[soilprofile2]{line_mod} function
}
\arguments{
\item{numberX}{A numeric vector that giving the number of points which should be added to the horizont line}

\item{name}{the name of horizonts}
}
\value{
This function returns a numeric id vector.
}
\description{
generates new X-values for the upper horizon limit
}
