% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_mod.R
\name{line_mod}
\alias{line_mod}
\title{creates a line as simple feature}
\usage{
line_mod(df_geom, line_attri, seed = 33)
}
\arguments{
\item{line_attri}{A datan frame with the information about the lines which divide the horizons. 
The following columns must exist:
\enumerate{
  \item The horizon Id must be given in a numeric column named "name"
  \item The number of X points, which should deviate the line from a straight line, must be given in a numeric column named: "numberX"
  \item The distance between the X points and the straight line must be given in a numeric column named: "sd". 
  The values are calculated using the \link[base]{rnorm} function. The mean value is given by the X-value of the straight horizon line. And the standard deviation can be selected individually. 
  \item The column named "sm" can be used as a logical value to specify whether the horizon transition should be smoothed. This is done with the function \link[smoothr]{smooth}.
}}

\item{seed}{For replicable examples a seed can be set here}

\item{df_geom_geom}{A long table with the columns named "name", "x" and "y" (Id and coordinates) for each horizont. For example created by the \link[soilprofile2]{cord_setting} function}
}
\value{
This function returns a linestring simple feature. Each line represents the upper horizontal limit of the corresponding id.
}
\description{
This is a simple function that creates a simple feature line from the horizont coordinates.
}
\examples{

## Example data
data_example <- data.frame(name = c(1, 2),
                          from1 = c(0,20),
                          to1 = c(20, 40)
)

##Coordination setting
cord_example <- cord_setting(data_example, plot_width = 3)

## Line attributes data frame:
lattri_example <- data.frame(name= unique(cord_example$name),
                            numberX = c(2, 10), 
                            sd = c(1,1),
                            sm = c(T, T)
)

## Apply the line_mod fuction
line_example <- line_mod(df = cord_example, line_attri = lattri_example)

## Plot simple feature geometry
plot(line_example$geometry)

}
