% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_default.R
\name{par_default}
\alias{par_default}
\title{par_default}
\usage{
par_default(polygon, size = 0.5, outline_col = "grey",
  size_grain = 1)
}
\arguments{
\item{polygon}{A simple feature file with plot information see \link[soilprofile2]{apply_texture}}

\item{size}{default size}

\item{outline_col}{dafault outline color of the horizont}

\item{size_grain}{the default is 1}
}
\value{
A unit multipolygon (sf) with one parID column
}
\description{
This function draws a set of randompolygons.
}
