% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_default.R
\name{par_default}
\alias{par_default}
\title{fill standard par}
\usage{
par_default(polygon)
}
\value{
A unit multipolygon (sf) with one parID column
}
\description{
This function draws a set of randompolygons.
}
