% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_default.R
\name{par_default}
\alias{par_default}
\title{par_default}
\usage{
par_default(polygon)
}
\arguments{
\item{polygon}{A Simple Feature geometry plot information see \link[soilprofile2]{apply_texture}}
}
\value{
A unit multipolygon (sf) with one parID column
}
\description{
This function draws a set of randompolygons.
}
