% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_import.R
\name{png_import}
\alias{png_import}
\title{imports png file an maks to the specific horizont position}
\usage{
png_import(file_path, horizont, smoothness, raster2polygon = T)
}
\arguments{
\item{file_path}{a numeric value between 1 and 6 calling the stored L-system}

\item{horizont}{A layer with the skeleton content as polygon (bei fehlen gibt es noch einen}

\item{smoothness}{passed to the smoothr function}

\item{raster2polygon}{an integer if the raster geometry shoud be transform to an vector polygon}
}
\description{
sourse sollte noch angepasst werden, evt dropbox oder db
}
