% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_import.R
\name{png_import}
\alias{png_import}
\title{imports png file an maks to the specific horizont position}
\usage{
png_import(path, mod_sf)
}
\arguments{
\item{path}{a numeric value between 1 and 6 calling the stored L-system}

\item{mod_sf}{A layer with the skeleton content as polygon (bei fehlen gibt es noch einen Fehler)}
}
\description{
sourse sollte noch angepasst werden, evt dropbox oder db
}
