% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_par_setting.R
\name{save_par_setting}
\alias{save_par_setting}
\title{save_horizont_fun}
\usage{
save_par_setting(name, variable, value, par_ID)
}
\arguments{
\item{name}{name of the function same to the horizontname}

\item{variable}{par variable e.g col}

\item{value}{value of variable e.g. blue}

\item{par_ID}{the "id" which defines the layer}
}
\value{
A unit sf layer with an parID column
}
\description{
This function save the new horizont function in the global list
}
