% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{soil skeleton}
\usage{
This function is used internally by the \link[soilprofile2]{line_mod} function. The corresponding Y values are created with the similar function \link[soilprofile2]{genY_fun}.
}
\arguments{
\item{X}{A numeric vector that giving the start and end of the horizont line}

\item{excited}{Logical value specifying whether to include an exclamation
point after the text}
}
\value{
This function returns 12 X-value. 10 of this are new generatet by the \link[base]{runif}  Funktion
}
\description{
generates new layer for the soil skeleton
}
