% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soil.R
\name{soil_theme}
\alias{soil_theme}
\title{wrapper function for plot}
\usage{
soil_theme()
}
\arguments{
\item{polygon}{A polygon-layer, which has been build by the \link[soilprofile2]{texture} function.}
}
\value{
A unit sf layer with an parID column
}
\description{
This function draws a Ah pattern, with the following  helper fuction \link[soilprofile2]{basic_point}
}
